+++ CHANGELOG.md
@@ -7,6 +7,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.1.6] - 2024-03-13
+
+### Added
+- Total VP for proposal votes
+
+### Changed
+- Ordering for proposal votes
+
 ## [0.1.5] - 2024-03-12
 
 ### Added